+++ src/base/config.ts
@@ -1,4 +1,0 @@
-export default interface Config {
-    prompt?: string;
-    ignore?: string[];
-}