+++ src/helpers/type-utils.ts
@@ -1,0 +1,3 @@
+export type ExtendWith<T, Extra> = {
+    [K in keyof T]: T[K] | Extra;
+};