+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2025-04-??
+
+### Added
+
+- Change from simple string testing to using glob (minimatch) for ignore lists
+
 ## [1.0.0] - 2025-04-02
 
 ### Added
@@ -9,8 +15,8 @@ All notable changes to this project will be documented in this file.
 - Initial CLI interface with `commander`
 - Recursive directory scanning
 - File tree formatting with emoji support
-- Support for `.config` files with `prompt` and `ignore`
+- Support for `ppllm.config.json` files with `prompt` and `ignore`
 - File content extraction with size/binary handling
-- Multi-language support (English, Polish)
+- Multi-language support
 - Output saving to file or stdout
 - Option parsing with robust error messages